+++ fastly/tls_custom_configuration_test.go
@@ -14,6 +14,38 @@ func TestClient_CustomTLSConfiguration(t *testing.T) {
 	var err error
 	conID := "TLS_CONFIGURATION_ID"
 
+	// Create
+	var ccon *CustomTLSConfiguration
+	Record(t, fixtureBase+"create", func(c *Client) {
+		ccon, err = c.CreateCustomTLSConfiguration(context.TODO(), &CreateCustomTLSConfigurationInput{
+			Name:          "My configuration",
+			HTTPProtocols: []string{"http/1.1", "http/2"},
+			TLSProtocols:  []string{"1.2", "1.3"},
+			Vipspace:      ToPointer("myvipspace"),
+			DefaultCertificate: &TLSCertificateRef{
+				ID: "1TTgJM5iiNcHi8V5cwCYk0",
+			},
+		})
+	})
+	if err != nil {
+		t.Fatal(err)
+	}
+	if conID != ccon.ID {
+		t.Errorf("bad ID: %q (%q)", conID, ccon.ID)
+	}
+	if len(ccon.TLS12CipherSuiteProfile) == 0 {
+		t.Errorf("expected TLS12CipherSuiteProfile to be populated")
+	}
+	if len(ccon.TLS13CipherSuiteProfile) == 0 {
+		t.Errorf("expected TLS13CipherSuiteProfile to be populated")
+	}
+	if ccon.DefaultCertificate == nil || ccon.DefaultCertificate.ID != "1TTgJM5iiNcHi8V5cwCYk0" {
+		t.Errorf("bad DefaultCertificate: %+v", ccon.DefaultCertificate)
+	}
+	if ccon.Vipspace == nil || *ccon.Vipspace != "myvipspace" {
+		t.Errorf("bad Vipspace: %+v", ccon.Vipspace)
+	}
+
 	// Get
 	var gcon *CustomTLSConfiguration
 	Record(t, fixtureBase+"get", func(c *Client) {
@@ -26,6 +58,18 @@ func TestClient_CustomTLSConfiguration(t *testing.T) {
 	}
 	if conID != gcon.ID {
 		t.Errorf("bad ID: %q (%q)", conID, gcon.ID)
+	}
+	if gcon.Vipspace == nil || *gcon.Vipspace != "myvipspace" {
+		t.Errorf("bad Vipspace: %v", gcon.Vipspace)
+	}
+	if len(gcon.TLS12CipherSuiteProfile) == 0 {
+		t.Errorf("expected TLS12CipherSuiteProfile to be populated")
+	}
+	if len(gcon.TLS13CipherSuiteProfile) == 0 {
+		t.Errorf("expected TLS13CipherSuiteProfile to be populated")
+	}
+	if gcon.DefaultCertificate == nil || gcon.DefaultCertificate.ID != "1TTgJM5iiNcHi8V5cwCYk0" {
+		t.Errorf("bad DefaultCertificate: %+v", gcon.DefaultCertificate)
 	}
 
 	// List
@@ -57,6 +101,16 @@ func TestClient_CustomTLSConfiguration(t *testing.T) {
 	}
 	if ucon.Name != newName {
 		t.Errorf("bad Name: %q (%q)", newName, ucon.Name)
+	}
+
+	// Delete
+	Record(t, fixtureBase+"delete", func(c *Client) {
+		err = c.DeleteCustomTLSConfiguration(context.TODO(), &DeleteCustomTLSConfigurationInput{
+			ID: conID,
+		})
+	})
+	if err != nil {
+		t.Fatal(err)
 	}
 }
 
@@ -116,6 +170,24 @@ func TestClient_UpdateCustomTLSConfiguration_validation(t *testing.T) {
 		ID: "CONFIGURATION_ID",
 	})
 	if !errors.Is(err, ErrMissingName) {
+		t.Errorf("bad error: %s", err)
+	}
+}
+
+func TestClient_CreateCustomTLSConfiguration_validation(t *testing.T) {
+	t.Parallel()
+
+	_, err := TestClient.CreateCustomTLSConfiguration(context.TODO(), &CreateCustomTLSConfigurationInput{})
+	if !errors.Is(err, ErrMissingHTTPProtocols) {
+		t.Errorf("bad error: %s", err)
+	}
+}
+
+func TestClient_DeleteCustomTLSConfiguration_validation(t *testing.T) {
+	t.Parallel()
+
+	err := TestClient.DeleteCustomTLSConfiguration(context.TODO(), &DeleteCustomTLSConfigurationInput{})
+	if !errors.Is(err, ErrMissingID) {
 		t.Errorf("bad error: %s", err)
 	}
 }